+++ src/modules/user/_test/data.js
@@ -1,0 +1,7 @@
+const arg = {
+    userInput: {
+        firstName: 'firstName3',
+        lastName: 'lastName3'
+    }
+};
+module.exports = {arg}
