+++ src/modules/user/_test/queries.js
@@ -1,0 +1,15 @@
+const userCreateM = `mutation UserCreate($userInput: UserItems) {
+  userCreate(userInput: $userInput) {
+    _id
+    firstName
+    lastName
+  }
+}`
+const userGetByIdQ = `query UserGetById($userId: ID!) {
+   userGetById(userId: $userId) {
+   _id
+   firstName
+   lastName
+   }
+ }`
+module.exports = {userCreateM, userGetByIdQ}
