+++ src/modules/user/_test/userCreate.spec.js
@@ -1,30 +1,16 @@
-const request = require('supertest')
-const {expect} = require('chai')
-const graphQLEndpoint = 'http://localhost:5000/graphql'
+const {expect} = require('chai');
+const{requestGql} = require('../../helper');
+const{userCreateM} = require('./queries');
+const{arg} = require('./data');
 
 describe('USER CREATE', () =>{
     describe('USER CREATE - POSITIVE', () =>{
         it('user create', (done) =>{
-
-            const arg = {
-                userInput: {
-                    firstName: 'firstName3',
-                    lastName: 'lastName3'
-                }
-            }
             const postData = {
-                query: `mutation UserCreate($userInput: UserItems) {
-                userCreate(userInput: $userInput) {
-                _id
-                 firstName
-                 lastName
-                 }
-                 }`,
-                variables: arg
-            }
-            request(graphQLEndpoint)
-                .post('/')
-                .send(postData)
+                query: userCreateM,
+                variables: arg,
+            };
+            requestGql(postData)
                 .expect(200)
                 .end((err, res) => {
                     if(err) return done(err);
@@ -32,13 +18,10 @@ describe('USER CREATE', () =>{
                     console.log('RESP BODY ===', respData)
                     expect(respData.userCreate.firstName).eq('firstName3')
                     expect(respData.userCreate.lastName).eq('lastName3')
-
                     done()
-
                 })
         })
     })
     describe('USER CREATE - NEGATIVE', () =>{
-
     })
 })
